WebVR;
- Add FORCE_FILESYSTEM
- Handle THREAD_ERROR event with noop case
- Remove hopefully redundant lovr.graphics.clear in errhand
- Finally use correct type signature for assimpFileTell

+++ src/data/modelData.c
@@ -201,7 +201,7 @@ static aiReturn assimpFileSeek(struct aiFile* assimpFile, size_t position, enum 
   return lovrFileSeek(file, position) ? aiReturn_FAILURE : aiReturn_SUCCESS;
 }
 
-static unsigned long assimpFileTell(struct aiFile* assimpFile) {
+static size_t assimpFileTell(struct aiFile* assimpFile) {
   File* file = (File*) assimpFile->UserData;
   return lovrFileTell(file);
 }
